requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
google-auth
google-api-python-client
google-cloud-documentai
python-dotenv
pydantic
